Remove MVKDescriptor.mm
There were no more descriptors in it, so it wasn't especially useful
