+++ MoltenVK/MoltenVK/GPUObjects/MVKPipeline.mm
@@ -19,6 +19,7 @@
 #include "MVKPipeline.h"
 #include "MVKCommandBuffer.h"
 #include "MVKInlineObjectConstructor.h"
+#include "MVKImage.h"
 #include "MVKFoundation.h"
 #include "MVKOSExtensions.h"
 #include "MVKStrings.h"
